requirements.txt
langchain_core
langchain_community
langchain
streamlit
python-dotenv